pyproject.toml
[project]
name = "claude-skills-automation"
version = "0.1.0"
description = "Claude Skillsを自動チューニングしてTerraform生成精度を向上させるシステム"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

